FIX Use 64-bit to track timestamps (#206)
to support dates beyond year 2106

This patch was done while reviewing potential year-2038 issues in openSUSE.

+++ modules/devices/battery.c
@@ -393,7 +393,7 @@ __scan_battery_apm(void)
 {
     FILE                *procapm;
     static char         *sremaining = NULL, *stotal = NULL;
-    static unsigned int  last_time = 0;
+    static long long int  last_time = 0;
     static int           percentage = 0;
     const  char         *ac_status[] = { "Battery",
                                          "AC Power",
